Work through, update, and document development setup
