+++ .gitignore
@@ -3,6 +3,7 @@
 /public/storage
 /storage/*.key
 /ssl/
+.idea/
 
 /vendor
 .env
